requirements.txt
flask==2.0.1
flask-restx==0.5.0
flask-sqlalchemy==2.5.1
flask-script==2.0.6
flask-migrate==3.0.1
gunicorn